refactor(graphics): split camera unit tests file (Issue #284 Phase6)
