+++ viewmodel/graphics/src/camera.rs
@@ -425,302 +425,5 @@ impl Default for Camera {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_camera_creation() {
-        let camera = Camera::new();
-
-        assert_eq!(camera.target, Vec3f::new(0.0, 0.0, 0.0));
-        assert_eq!(camera.distance, 5.0);
-        assert!(camera.rotation.is_unit());
-    }
-
-    #[test]
-    fn test_camera_rotation() {
-        let mut camera = Camera::new();
-        let initial_rotation = camera.rotation;
-
-        camera.rotate(1.0, 0.0);
-
-        // 回転が変化していることを確認
-        assert_ne!(camera.rotation.w(), initial_rotation.w());
-    }
-
-    #[test]
-    fn test_camera_pan() {
-        let mut camera = Camera::new();
-        let initial_target = camera.target;
-
-        camera.pan(1.0, 0.0);
-
-        // ターゲット位置が変化していることを確認
-        assert_ne!(camera.target, initial_target);
-    }
-
-    #[test]
-    fn test_camera_zoom() {
-        let mut camera = Camera::new();
-        let initial_distance = camera.distance;
-
-        camera.zoom(1.0, 1.0); // 拡大
-
-        // 距離が変化していることを確認
-        assert_ne!(camera.distance, initial_distance);
-        assert!(camera.distance >= 0.1); // 最小制限確認
-        assert!(camera.distance <= 50.0); // 最大制限確認
-    }
-
-    #[test]
-    fn test_camera_slerp() {
-        let camera1 = Camera::new();
-        let mut camera2 = Camera::new();
-        camera2.rotate(1.0, 0.0); // 少し回転
-
-        let interpolated = camera1.slerp_to(&camera2, 0.5).unwrap();
-
-        // 補間された回転が両者の中間にあることを確認
-        assert!(interpolated.rotation.dot(&camera1.rotation) > 0.5);
-        assert!(interpolated.rotation.dot(&camera2.rotation) > 0.5);
-    }
-
-    #[test]
-    fn test_quaternion_to_matrix() {
-        let q = Quaternionf::identity();
-        let matrix = crate::camera_math::quaternion_to_matrix(&q);
-
-        // 単位行列のテスト
-        assert!((matrix[0][0] - 1.0).abs() < 1e-6);
-        assert!((matrix[1][1] - 1.0).abs() < 1e-6);
-        assert!((matrix[2][2] - 1.0).abs() < 1e-6);
-        assert!((matrix[3][3] - 1.0).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_fit_to_small_mesh() {
-        let mut camera = Camera::new();
-
-        // 1mm立方体のテスト
-        let min_bounds = Vec3f::new(-0.0005, -0.0005, -0.0005); // -0.5mm
-        let max_bounds = Vec3f::new(0.0005, 0.0005, 0.0005); // +0.5mm
-
-        camera.fit_to_small_mesh(min_bounds, max_bounds);
-
-        // ターゲットが中心になっていることを確認
-        assert_eq!(camera.target, Vec3f::new(0.0, 0.0, 0.0));
-
-        // 距離が適切に設定されていることを確認（対角線の10倍程度）
-        let diagonal = (3.0_f32 * 0.001_f32.powi(2)).sqrt(); // ≈ 0.00173
-        assert!(camera.distance > diagonal * 5.0); // 最低5倍
-        assert!(camera.distance < diagonal * 20.0); // 最大20倍
-
-        // 回転が設定されていることを確認
-        assert_ne!(camera.rotation, Quaternionf::identity());
-    }
-
-    #[test]
-    fn test_projection_matrix_near_far() {
-        let mut camera = Camera::new();
-        camera.distance = 0.01; // 非常に近い距離
-
-        let proj = camera.projection_matrix(1.0);
-
-        // プロジェクション行列が生成されることを確認
-        assert!(!proj[0][0].is_nan());
-        assert!(!proj[1][1].is_nan());
-        assert!(!proj[2][2].is_nan());
-        assert!(!proj[3][2].is_nan());
-    }
-
-    #[test]
-    fn test_projection_modes() {
-        let mut camera = Camera::new();
-
-        // デフォルトは透視投影
-        assert_eq!(camera.projection_mode, ProjectionMode::Perspective);
-
-        // 平行投影に切り替え
-        camera.set_projection_mode(ProjectionMode::Orthographic);
-        assert_eq!(camera.projection_mode, ProjectionMode::Orthographic);
-
-        // 両方の投影モードで行列が生成されることを確認
-        let perspective_proj = {
-            camera.set_projection_mode(ProjectionMode::Perspective);
-            camera.projection_matrix(1.0)
-        };
-
-        let orthographic_proj = {
-            camera.set_projection_mode(ProjectionMode::Orthographic);
-            camera.projection_matrix(1.0)
-        };
-
-        // 異なる投影行列が生成されることを確認
-        assert_ne!(perspective_proj[0][0], orthographic_proj[0][0]);
-        assert_ne!(perspective_proj[2][2], orthographic_proj[2][2]);
-    }
-
-    #[test]
-    fn test_orthographic_camera_creation() {
-        let ortho_camera = Camera::new_orthographic();
-        assert_eq!(ortho_camera.projection_mode, ProjectionMode::Orthographic);
-
-        let persp_camera = Camera::new();
-        assert_eq!(persp_camera.projection_mode, ProjectionMode::Perspective);
-    }
-
-    #[test]
-    fn test_isometric_camera_creation() {
-        let isometric_camera = Camera::new_isometric();
-        assert_eq!(
-            isometric_camera.projection_mode,
-            ProjectionMode::Orthographic
-        );
-        assert_eq!(isometric_camera.zoom, 1.0);
-
-        // アイソメトリック回転が適用されているかチェック
-        assert_ne!(isometric_camera.rotation, Quaternionf::identity());
-
-        // 正規化されているかチェック（クォータニオンのノルムは1）
-        let norm = (isometric_camera.rotation.w().powi(2)
-            + isometric_camera.rotation.x().powi(2)
-            + isometric_camera.rotation.y().powi(2)
-            + isometric_camera.rotation.z().powi(2))
-        .sqrt();
-        assert!((norm - 1.0).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_camera_reset_functions() {
-        let mut camera = Camera::new();
-        camera.distance = 0.1; // 非常に近い距離
-        camera.target = Vec3f::new(5.0, 5.0, 5.0);
-
-        // 基本リセット
-        camera.reset();
-        assert_eq!(camera.distance, 5.0);
-        assert_eq!(camera.target, Vec3f::new(0.0, 0.0, 0.0));
-        assert_eq!(camera.zoom, 1.0);
-
-        // 最小距離確保
-        camera.distance = 0.05; // 危険な近距離
-        camera.ensure_minimum_distance();
-        assert!(camera.distance >= 1.0);
-
-        // 安全な視点リセット
-        let min_bounds = Vec3f::new(-0.5, -0.5, -0.5);
-        let max_bounds = Vec3f::new(0.5, 0.5, 0.5);
-        camera.reset_to_safe_view(min_bounds, max_bounds);
-        assert!(camera.distance >= 2.0); // 安全な距離
-        assert_eq!(camera.target, Vec3f::new(0.0, 0.0, 0.0)); // 中心
-    }
-
-    // ============================================================================
-    // Issue #242 テストスイート：カメラ操作の4つの問題の検証
-    // ============================================================================
-
-    #[test]
-    fn test_issue_242_arcball_sphere_mapping() {
-        // Issue #242 Phase 1: 回転操作の球面定義
-        // 画面座標を単位球面上の点にマッピング
-
-        // ビューポート中心のクリック（Z軸正方向）
-        let center = Camera::project_on_sphere(400.0, 300.0, 800.0, 600.0);
-        assert!((center.z() - 1.0).abs() < 0.01, "中心クリック時はZ≈1");
-        assert!(center.x().abs() < 0.01, "中心クリック時はX≈0");
-        assert!(center.y().abs() < 0.01, "中心クリック時はY≈0");
-
-        // ビューポート左端のクリック
-        let left = Camera::project_on_sphere(0.0, 300.0, 800.0, 600.0);
-        assert!(left.x() < 0.0, "左端クリック時はX<0");
-
-        // ビューポート右端のクリック
-        let right = Camera::project_on_sphere(800.0, 300.0, 800.0, 600.0);
-        assert!(right.x() > 0.0, "右端クリック時はX>0");
-
-        // 全ての点が単位球面上にあることを確認（正規化済み）
-        for point in [center, left, right].iter() {
-            let magnitude = (point.x().powi(2) + point.y().powi(2) + point.z().powi(2)).sqrt();
-            assert!((magnitude - 1.0).abs() < 0.01, "全点が正規化済み");
-        }
-    }
-
-    #[test]
-    fn test_issue_242_arcball_rotation_from_sphere_points() {
-        // Issue #242 Phase 1: 球面上の2点から回転を計算
-
-        // 同じ点 → 回転なし
-        let p1 = Vec3f::new(1.0, 0.0, 0.0);
-        let q_identity = Camera::compute_rotation_from_sphere_points(p1, p1);
-        let identity = Quaternionf::identity();
-        // 恒等元の確認：(w, x, y, z) = (1, 0, 0, 0)
-        assert!(
-            (q_identity.w() - identity.w()).abs() < 0.001
-                && (q_identity.x() - identity.x()).abs() < 0.001
-                && (q_identity.y() - identity.y()).abs() < 0.001
-                && (q_identity.z() - identity.z()).abs() < 0.001,
-            "同じ点からの回転は恒等元"
-        );
-
-        // 90度の角度
-        let p_start = Vec3f::new(1.0, 0.0, 0.0).normalize().unwrap();
-        let p_end = Vec3f::new(0.0, 1.0, 0.0).normalize().unwrap();
-        let q_90 = Camera::compute_rotation_from_sphere_points(p_start, p_end);
-        // 回転が生成されている（恒等元ではない）
-        let is_not_identity = (q_90.w() - identity.w()).abs() > 0.01
-            || (q_90.x() - identity.x()).abs() > 0.01
-            || (q_90.y() - identity.y()).abs() > 0.01
-            || (q_90.z() - identity.z()).abs() > 0.01;
-        assert!(is_not_identity, "異なる点からは回転が生成される");
-
-        // 逆方向 → 逆回転
-        let q_rev = Camera::compute_rotation_from_sphere_points(p_end, p_start);
-        let is_rev_not_identity = (q_rev.w() - identity.w()).abs() > 0.01
-            || (q_rev.x() - identity.x()).abs() > 0.01
-            || (q_rev.y() - identity.y()).abs() > 0.01
-            || (q_rev.z() - identity.z()).abs() > 0.01;
-        assert!(is_rev_not_identity, "逆方向も回転を生成");
-    }
-
-    #[test]
-    fn test_issue_242_zoom_magnitude_sensitivity() {
-        // Issue #242: ズーム反応改善
-        // マウス移動の大きさ（ノルム）で感度を統一
-
-        let mut camera_vertical = Camera::new();
-        let mut camera_horizontal = Camera::new();
-        let mut camera_diagonal = Camera::new();
-
-        let initial_distance = camera_vertical.distance;
-
-        // 縦方向のみのドラッグ: (0, 10) → magnitude = 10
-        camera_vertical.zoom(0.0, 10.0);
-
-        // 横方向のみのドラッグ: (10, 0) → magnitude = 10
-        camera_horizontal.zoom(10.0, 0.0);
-
-        // 斜めドラッグ: (7, 7) → magnitude ≈ 10
-        camera_diagonal.zoom(7.0, 7.0);
-
-        // 3つのズーム量が異なることを確認
-        // （方向によって拡大/縮小が変わるため、エラー許容度を広くする）
-        let dist_v = camera_vertical.distance;
-        let dist_h = camera_horizontal.distance;
-        let dist_d = camera_diagonal.distance;
-
-        println!(
-            "ズーム結果: vertical={:.3}, horizontal={:.3}, diagonal={:.3}",
-            dist_v, dist_h, dist_d
-        );
-
-        // 3つの値が大異なるわけではなく、移動量の大きさに応じて変化
-        assert!(
-            (dist_v - initial_distance).abs() > 0.001,
-            "縦方向ズームが有効"
-        );
-        assert!(
-            (dist_h - initial_distance).abs() > 0.001,
-            "横方向ズームが有効"
-        );
-    }
-}
+#[path = "camera_tests.rs"]
+mod tests;
